# server/src/interfaces/http/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a ecommerce project",
    "version": "1.0.0",
    "title": "iPragmatech Ecommerce",
    "termsOfService": "",
    "contact": {
      "email": "kapil.jain@ipragmatech.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "ipragmatech-ecommmerce.herokuapp.com",
  "basePath": "/api",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/store/order": {
      "post": {
        "tags": [
          "store"
        ],
        "summary": "Place an order for a products",
        "description": "",
        "operationId": "placeOrder",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "order placed for purchasing the product",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Cart"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "400": {
            "description": "Invalid Order"
          }
        }
      }
    },
    "/store/order/{orderId}": {
      "get": {
        "tags": [
          "store"
        ],
        "summary": "Find purchase order by ID",
        "description": "For valid response try integer IDs with value >= 1 and <= 10. Other values will generated exceptions",
        "operationId": "getOrderById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "ID of order that needs to be fetched",
            "required": true,
            "type": "integer",
            "maximum": 10,
            "minimum": 1,
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Order not found"
          }
        }
      },
      "put": {
        "tags": [
          "store"
        ],
        "summary": "Find purchase order by ID",
        "description": "For valid response try integer IDs with value >= 1 and <= 10. Other values will generated exceptions",
        "operationId": "updateOrderById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "ID of order that needs to be fetched",
            "required": true,
            "type": "integer",
            "maximum": 10,
            "minimum": 1,
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Update order with status",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    },
...
    "/store/cart": {
      "post": {
        "tags": [
          "store"
        ],
        "summary": "Add item to cart with default quantity as 1",
        "description": "",
        "operationId": "addToCart",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Add product in the cart",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Cart"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Cart"
            }
          },
          "400": {
            "description": "Invalid Cart"
          }
        }
      },
      "get": {
        "tags": [
          "store"
        ],
        "summary": "Get items from cart",
        "description": "",
        "operationId": "getCart",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Cart"
              }
            }
          },
          "400": {
            "description": "Invalid Cart"
          }
        }
      }
    },
    "/store/cart/:itemId": {
      "put": {
        "tags": [
          "store"
        ],
        "summary": "Update item to cart with quantity",
        "description": "",
        "operationId": "updateCart",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Update product in the cart",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CartItemQuantity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Cart"
            }
          },
          "400": {
            "description": "Invalid Cart"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Order": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "integer",
          "format": "int64"
        },
        "customerId": {
          "type": "integer",
          "format": "int64"
        },
        "totalAmount": {
          "type": "integer",
          "format": "int32"
        },
        "shipDate": {
          "type": "string",
          "format": "date-time"
        },
        "createDate": {
          "type": "string",
          "format": "date-time"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "comments": {
          "type": "string"
        },
        "authCode": {
          "type": "string"
        },
        "reference": {
          "type": "string"
        },
        "shippingId": {
          "type": "integer"
        },
        "taxId": {
          "type": "integer"
        }
      }
    },
...
    "OrderStatus": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32",
          "example": "Status 1 for success payment, 2 for cancel payment"
        },
        "comments": {
          "type": "string"
        },
        "authCode": {
          "type": "string",
          "example": "PaymentId recieved from the payment gateway"
        },
        "reference": {
          "type": "string",
          "example": "Json Object received from the payment gateway"
        }
      }
    },
    "Cart": {
      "type": "object",
      "properties": {
        "itemId": {
          "type": "integer",
          "format": "int64"
        },
        "cartId": {
          "type": "integer",
          "format": "int64"
        },
        "productId": {
          "type": "integer",
          "format": "int64"
        },
        "attributes": {
          "type": "string"
        },
        "quantity": {
          "type": "integer"
        },
        "buyNow": {
          "type": "integer"
        },
        "createdDate": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "CartItemQuantity": {
      "type": "object",
      "properties": {
        "quantity": {
          "type": "integer"
        }
      }
    },
...
  }
}